% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysumfunction.R
\name{mysumfunction}
\alias{mysumfunction}
\title{Title A number plus 1 function}
\usage{
mysumfunction(s)
}
\arguments{
\item{s}{\itemize{
\item a numeric value
}}
}
\value{
output  input number plus one
}
\description{
Title A number plus 1 function
}
\examples{
mysum(3)
}
