% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysumfunction.R
\name{sumfunction_5}
\alias{sumfunction_5}
\title{Title}
\usage{
sumfunction_5(s)
}
\arguments{
\item{s}{\itemize{
\item a numeric value
}}
}
\value{
output  input number plus one
}
\description{
Title
}
\examples{
sumfunction_5(3)
}
